#114 Aditional data to $auth.authenticate
Made possible that the $auth.authenticate function can receive a second parameter (Object) that will be passed all the way down the oauth request pipeline (either OAuth1 or OAuth2) to be sent the the server's final post.
ex. In a Facebook login, after the popup has closed, it will be object will be serialized along with the code, clientId and redirectUri and sent to the server